+++ Items/Tokens/tier2/TopazToken.cs
@@ -36,28 +36,28 @@ namespace prefixtest.Items.Tokens.tier2
 		switch (x)
 		{
 				case 0:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<earthcharm>(), 1);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<earthcharm>(), 1);
 						break;
 				case 1:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<shotgunshot>(), 999);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<shotgunshot>(), 999);
 						break;
 				case 2:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<gemstave>(), 1);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<gemstave>(), 1);
 						break;
 				case 3:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<heavycrossbow>(), 1);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<heavycrossbow>(), 1);
 						break;
 				case 4:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<rimefrost>(), 1);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<rimefrost>(), 1);
 						break;
 				case 5:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<shotgunblade>(), 1);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<shotgunblade>(), 1);
 						break;
 				case 6:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<vampireknife>(), 1);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<vampireknife>(), 1);
 						break;
 				case 7:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<soulofchance>(), Main.rand.Next(2, 5));
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<soulofchance>(), Main.rand.Next(2, 5));
 						break;
 		}
 		return true;
