+++ Items/Tokens/tier4/EmeraldToken.cs
@@ -35,25 +35,25 @@ namespace prefixtest.Items.Tokens.tier4
 		switch (x)
 		{
 				case 0:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<blazereap4>(), 1);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<blazereap4>(), 1);
 						break;
 				case 1:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<starshot>(), 999);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<starshot>(), 999);
 						break;
 				case 2:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<dreamshatter>(), 1);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<dreamshatter>(), 1);
 						break;
 				case 3:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<infinityedge>(), 1);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<infinityedge>(), 1);
 						break;
 				case 4:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<quazar>(), 1);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<quazar>(), 1);
 						break;
 				case 5:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<theecho>(), 1);
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<theecho>(), 1);
 						break;
 				case 6:
-						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<soulofchance>(), Main.rand.Next(2, 5));
+						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<soulofchance>(), Main.rand.Next(2, 5));
 						break;
 		}
 		return true;
